+++ Samples~/MVVM/Runtime/MyAppBuilder.cs
@@ -1,10 +1,21 @@
 using Unity.AppUI.MVVM;
 using UnityEngine;
+using UnityEngine.UIElements;
 
 namespace Unity.AppUI.Samples.MVVM
 {
     public class MyAppBuilder : UIToolkitAppBuilder<MyApp>
     {
+        void Awake()
+        {
+#if ENABLE_PANEL_RENDERER
+            panelRenderer = GetComponent<PanelRenderer>();
+            if (panelRenderer)
+                return;
+#endif
+            uiDocument = GetComponent<UIDocument>();
+        }
+
         protected override void OnAppInitialized(MyApp app)
         {
             base.OnAppInitialized(app);